requirements.txt
langchain==1.0.2
langchain-openai==1.0.1
langchain-anthropic>=0.1.0
langchain-mcp-adapters>=0.1.0
fastmcp==2.12.5
starlette
uvicorn
python-dotenv
requests
toon-format